requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
chromadb
PyPDF2
python-docx
python-dotenv
openai
